requirements.txt
ipympl
u8darts[all]
ipywidgets
yfinance
pandas-ta
dataclasses